refactor: favorite 도메인 네이밍 개선
