+++ src/main/java/org/example/pedia_777/domain/favorite/service/FavoriteService.java
@@ -3,10 +3,10 @@ package org.example.pedia_777.domain.favorite.service;
 import lombok.RequiredArgsConstructor;
 import org.example.pedia_777.common.dto.PageResponse;
 import org.example.pedia_777.common.exception.BusinessException;
-import org.example.pedia_777.domain.favorite.code.FavoriteErrorCode;
 import org.example.pedia_777.domain.favorite.dto.response.FavoriteAddResponse;
 import org.example.pedia_777.domain.favorite.dto.response.FavoriteMovieResponse;
 import org.example.pedia_777.domain.favorite.entity.Favorite;
+import org.example.pedia_777.domain.favorite.error.FavoriteErrorCode;
 import org.example.pedia_777.domain.favorite.repository.FavoriteRepository;
 import org.example.pedia_777.domain.member.entity.Member;
 import org.example.pedia_777.domain.member.service.MemberServiceApi;
